# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: College Management System API for Tamil Nadu
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hostels:
    get:
      operationId: listHostels
      tags: [hostels]
      summary: List hostels
      responses:
        "200":
          description: Hostel list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hostel"
    post:
      operationId: createHostel
      tags: [hostels]
      summary: Create hostel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostelBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hostel"
  /hostels/{id}:
    get:
      operationId: getHostel
      tags: [hostels]
      summary: Get hostel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hostel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hostel"
    patch:
      operationId: updateHostel
      tags: [hostels]
      summary: Update hostel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHostelBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hostel"
    delete:
      operationId: deleteHostel
      tags: [hostels]
      summary: Delete hostel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /hostel-rooms:
    get:
      operationId: listHostelRooms
      tags: [hostels]
      summary: List hostel rooms
      parameters:
        - name: hostelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Room list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostelRoom"
    post:
      operationId: createHostelRoom
      tags: [hostels]
      summary: Create hostel room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostelRoomBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostelRoom"
  /hostel-rooms/{id}:
    patch:
      operationId: updateHostelRoom
      tags: [hostels]
      summary: Update hostel room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHostelRoomBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostelRoom"
    delete:
      operationId: deleteHostelRoom
      tags: [hostels]
      summary: Delete hostel room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /hostel-allocations:
    get:
      operationId: listHostelAllocations
      tags: [hostels]
      summary: List hostel allocations
      parameters:
        - name: hostelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Allocation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostelAllocation"
    post:
      operationId: createHostelAllocation
      tags: [hostels]
      summary: Create hostel allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostelAllocationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostelAllocation"
  /hostel-allocations/{id}:
    patch:
      operationId: updateHostelAllocation
      tags: [hostels]
      summary: Update hostel allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHostelAllocationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostelAllocation"
  /hostel-complaints:
    get:
      operationId: listHostelComplaints
      tags: [hostels]
      summary: List hostel complaints
      parameters:
        - name: hostelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Complaint list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostelComplaint"
    post:
      operationId: createHostelComplaint
      tags: [hostels]
      summary: Create hostel complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostelComplaintBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostelComplaint"
  /hostel-complaints/{id}:
    patch:
      operationId: updateHostelComplaint
      tags: [hostels]
      summary: Update hostel complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHostelComplaintBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostelComplaint"

  # ── Transport ──
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /event-participants:
    get:
      operationId: listEventParticipants
      tags: [events]
      summary: List event participants
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Participant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventParticipant"
    post:
      operationId: createEventParticipant
      tags: [events]
      summary: Register participant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventParticipantBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventParticipant"
  /event-participants/{id}:
    patch:
      operationId: updateEventParticipant
      tags: [events]
      summary: Update participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventParticipantBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventParticipant"
    delete:
      operationId: deleteEventParticipant
      tags: [events]
      summary: Delete participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── Communications ──
...
components:
  schemas:
...
    Hostel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        totalBlocks:
          type: integer
        totalRooms:
          type: integer
        wardenName:
          type: string
          nullable: true
        wardenPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        facilities:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, totalBlocks, totalRooms, status, createdAt]
    CreateHostelBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        totalBlocks:
          type: integer
        totalRooms:
          type: integer
        wardenName:
          type: string
        wardenPhone:
          type: string
        address:
          type: string
        facilities:
          type: string
        status:
          type: string
      required: [name, type]
    UpdateHostelBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        totalBlocks:
          type: integer
        totalRooms:
          type: integer
        wardenName:
          type: string
        wardenPhone:
          type: string
        address:
          type: string
        facilities:
          type: string
        status:
          type: string
    HostelRoom:
      type: object
      properties:
        id:
          type: integer
        hostelId:
          type: integer
        roomNumber:
          type: string
        floor:
          type: integer
        block:
          type: string
          nullable: true
        roomType:
          type: string
        capacity:
          type: integer
        occupancy:
          type: integer
        amenities:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, hostelId, roomNumber, floor, roomType, capacity, occupancy, status, createdAt]
    CreateHostelRoomBody:
      type: object
      properties:
        hostelId:
          type: integer
        roomNumber:
          type: string
        floor:
          type: integer
        block:
          type: string
        roomType:
          type: string
        capacity:
          type: integer
        amenities:
          type: string
        status:
          type: string
      required: [hostelId, roomNumber, roomType]
    UpdateHostelRoomBody:
      type: object
      properties:
        roomNumber:
          type: string
        floor:
          type: integer
        block:
          type: string
        roomType:
          type: string
        capacity:
          type: integer
        occupancy:
          type: integer
        amenities:
          type: string
        status:
          type: string
    HostelAllocation:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        hostelId:
          type: integer
        roomId:
          type: integer
        academicYear:
          type: string
        allocationDate:
          type: string
        vacatingDate:
          type: string
          nullable: true
        messType:
          type: string
          nullable: true
        status:
          type: string
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, hostelId, roomId, academicYear, allocationDate, status, createdAt]
    CreateHostelAllocationBody:
      type: object
      properties:
        studentId:
          type: integer
        hostelId:
          type: integer
        roomId:
          type: integer
        academicYear:
          type: string
        allocationDate:
          type: string
        messType:
          type: string
        status:
          type: string
        remarks:
          type: string
      required: [studentId, hostelId, roomId, academicYear, allocationDate]
    UpdateHostelAllocationBody:
      type: object
      properties:
        vacatingDate:
          type: string
        messType:
          type: string
        status:
          type: string
        remarks:
          type: string
    HostelComplaint:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        hostelId:
          type: integer
        category:
          type: string
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        resolvedDate:
          type: string
          nullable: true
        resolution:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, hostelId, category, subject, description, priority, status, createdAt]
    CreateHostelComplaintBody:
      type: object
      properties:
        studentId:
          type: integer
        hostelId:
          type: integer
        category:
          type: string
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
      required: [studentId, hostelId, category, subject, description]
    UpdateHostelComplaintBody:
      type: object
      properties:
        status:
          type: string
        resolution:
          type: string
        resolvedDate:
          type: string
        priority:
          type: string

    # ── Transport Schemas ──
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        departmentId:
          type: integer
          nullable: true
        venue:
          type: string
          nullable: true
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        coordinatorName:
          type: string
          nullable: true
        coordinatorPhone:
          type: string
          nullable: true
        maxParticipants:
          type: integer
          nullable: true
        budget:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, startDate, status, createdAt]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        description:
          type: string
        departmentId:
          type: integer
        venue:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coordinatorName:
          type: string
        coordinatorPhone:
          type: string
        maxParticipants:
          type: integer
        budget:
          type: string
        status:
          type: string
      required: [title, type, startDate]
    UpdateEventBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        description:
          type: string
        departmentId:
          type: integer
        venue:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        coordinatorName:
          type: string
        coordinatorPhone:
          type: string
        maxParticipants:
          type: integer
        budget:
          type: string
        status:
          type: string
    EventParticipant:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        studentId:
          type: integer
        role:
          type: string
        registrationDate:
          type: string
        achievement:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, studentId, role, registrationDate, createdAt]
    CreateEventParticipantBody:
      type: object
      properties:
        eventId:
          type: integer
        studentId:
          type: integer
        role:
          type: string
        registrationDate:
          type: string
        achievement:
          type: string
        remarks:
          type: string
      required: [eventId, studentId, registrationDate]
    UpdateEventParticipantBody:
      type: object
      properties:
        role:
          type: string
        achievement:
          type: string
        remarks:
          type: string

    # ── Communication Schemas ──
...