# lib/api-zod/src/generated/types/createEventParticipantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Management System API for Tamil Nadu
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventParticipantBody {
  eventId: number;
  studentId: number;
  role?: string;
  registrationDate: string;
  achievement?: string;
  remarks?: string;
}