# lib/api-zod/src/generated/types/hostelAllocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Management System API for Tamil Nadu
 * OpenAPI spec version: 0.1.0
 */

export interface HostelAllocation {
  id: number;
  studentId: number;
  hostelId: number;
  roomId: number;
  academicYear: string;
  allocationDate: string;
  vacatingDate?: string | null;
  messType?: string | null;
  status: string;
  remarks?: string | null;
  createdAt: Date;
}