# lib/api-zod/src/generated/types/hostelRoom.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Management System API for Tamil Nadu
 * OpenAPI spec version: 0.1.0
 */

export interface HostelRoom {
  id: number;
  hostelId: number;
  roomNumber: string;
  floor: number;
  block?: string | null;
  roomType: string;
  capacity: number;
  occupancy: number;
  amenities?: string | null;
  status: string;
  createdAt: Date;
}